refactor(core): decouple cache from lint, expose skipRules + hasFix API
Walks back the cache-aware lint() introduced in 2284d7c. The cache
file format and lifecycle are CLI concerns; core's job is to run
rules. Future cache changes (BuilderProgram-based invalidation,
schema bumps) shouldn't churn this layer.

API changes on createLinter():

- `lint(fileName, options?: { skipRules?: ReadonlySet<string> })`
  Replaces the cache-mutation overload. Caller passes rule IDs to
  skip; core just doesn't run them. Caller is responsible for merging
  cached output back into the diagnostic list.
- `hasFixForDiagnostic(fileName, diagnostic): boolean`
  New per-diagnostic fix accessor. The CLI cache layer uses this to
  snapshot the `hasFix` flag for a rule's cache entry without reaching
  into core's internal `diagnostic2Fixes` map.
- Removed types: `RuleCache`, `FileLintCache`. Both move to the CLI
  cache module in the next commit.
- Removed the cache-write side of `report()`. Reports now only update
  the in-memory lint result; serialization for persistence is handled
  outside core.

What stays in core:
- The runtime probe + sticky `typeAwareRules` set + `getTypeAwareRules()`
  exposed for the CLI to persist into `ruleModes`.
- `initialTypeAwareRules` constructor param so a session starting cold
  can preseed classification from the cache file.

Tests:
- packages/core/test/cache-layer1.test.ts removed — that suite tested
  cache mutation through lint(), which is no longer core's job. The
  same scenarios are reformulated against the CLI cache layer in the
  next commit.
- packages/core/test/skip-rules.test.ts added (9 checks): rule in
  skipRules doesn't run, only specified rule is skipped, lint without
  options runs everything (back-compat), seeded classification persists
  through skipping, hasFixForDiagnostic returns true only for diags
  with registered fixes.

Existing probe.test.ts and builder-program-poc.test.ts pass unchanged.
