+++ .github/workflows/test.yml
@@ -66,9 +66,10 @@ jobs:
       - name: Core BuilderProgram POC
         run: node packages/core/test/builder-program-poc.test.js
 
-      # Core: layer 1 cache (per-file, per-rule, syntactic-only).
-      - name: Core layer 1 cache
-        run: node packages/core/test/cache-layer1.test.js
+      # Core: skip-rules option + hasFixForDiagnostic API (the surface
+      # the CLI cache layer uses).
+      - name: Core skip-rules
+        run: node packages/core/test/skip-rules.test.js
 
       # CLI: cache file load / save / atomic write / version key.
       - name: CLI cache module
