+++ packages/core/index.ts
@@ -4,23 +4,6 @@ import type * as ts from 'typescript';
 import path = require('path');
 import minimatch = require('minimatch');
 
-// Per-file diagnostic cache. Keyed by ruleId — each entry holds the
-// rule's last-known result (with `hasFix` so the CLI can decide whether
-// to surface `--fix` even on a cache hit). The CLI handles file-level
-// invalidation (mtime check) and cross-session persistence; this type
-// is just the in-memory shape core reads from and writes to.
-//
-// Diagnostics stored here have `file` cleared (`undefined as any`) —
-// they're rehydrated with the current `ts.SourceFile` on cache hit.
-// The CLI's serialization layer drops / restores them across the
-// JSON boundary.
-export interface RuleCache {
-	hasFix: boolean;
-	diagnostics: ts.DiagnosticWithLocation[];
-}
-
-export type FileLintCache = Record</* ruleId */ string, RuleCache>;
-
 export type Linter = ReturnType<typeof createLinter>;
 
 export function createLinter(
@@ -68,8 +51,18 @@ export function createLinter(
 	const typeAwareRules = new Set<string>(initialTypeAwareRules ?? []);
 
 	return {
-		lint(fileName: string, fileCache?: FileLintCache): ts.DiagnosticWithLocation[] {
+		// `options.skipRules`: rule IDs the caller already has cached
+		// results for. Core just doesn't run them — the caller is
+		// responsible for merging cached output into the final diagnostic
+		// list. Decoupling cache lifecycle from core means future cache
+		// changes (BuilderProgram-based invalidation, schema bumps) live
+		// in the CLI without churning this layer.
+		lint(
+			fileName: string,
+			options?: { skipRules?: ReadonlySet<string> },
+		): ts.DiagnosticWithLocation[] {
 			let currentRuleId: string;
+			const skipRules = options?.skipRules;
 
 			const rules = getRulesForFile(fileName);
 			const token = ctx.languageServiceHost.getCancellationToken?.();
@@ -99,24 +92,7 @@ export function createLinter(
 
 				currentRuleId = ruleId;
 
-				const ruleCache = fileCache?.[currentRuleId];
-				// Cache hit only when:
-				//   - cache has an entry for this rule
-				//   - rule has not been classified type-aware
-				// For type-aware rules we never trust the cache, regardless
-				// of whether the entry was written this session or carried
-				// over from before classification.
-				if (ruleCache && !typeAwareRules.has(currentRuleId)) {
-					for (const cached of ruleCache.diagnostics) {
-						lintResult[1].set({
-							...cached,
-							file: rulesContext.file,
-							relatedInformation: cached.relatedInformation?.map(info => ({
-								...info,
-								file: info.file ? program.getSourceFile(info.file.fileName) : undefined,
-							})),
-						}, []);
-					}
+				if (skipRules?.has(currentRuleId)) {
 					continue;
 				}
 
@@ -134,25 +110,6 @@ export function createLinter(
 				}
 				if (touchedProgram) {
 					typeAwareRules.add(currentRuleId);
-				}
-
-				if (fileCache) {
-					if (typeAwareRules.has(currentRuleId)) {
-						// Discard any entry — could have been written by
-						// `report()` during the run before the program access,
-						// or stale from a prior session that didn't classify
-						// this rule.
-						delete fileCache[currentRuleId];
-					}
-					else {
-						fileCache[currentRuleId] ??= { hasFix: false, diagnostics: [] };
-						for (const [_, fixes] of lintResult[1]) {
-							if (fixes.length) {
-								fileCache[currentRuleId].hasFix = true;
-								break;
-							}
-						}
-					}
 				}
 			}
 
@@ -195,23 +152,6 @@ export function createLinter(
 				};
 				let location: [Error, number] = [new Error(), 1];
 				let relatedInformation: ts.DiagnosticRelatedInformation[] | undefined;
-
-				// Push a serialization-friendly twin of the diagnostic into
-				// the cache as the rule reports. If the rule turns out to be
-				// type-aware (touchedProgram flips later in the same call),
-				// the post-rule cleanup deletes this entry — the cache only
-				// retains entries for rules confirmed syntactic.
-				if (fileCache) {
-					fileCache[currentRuleId] ??= { hasFix: false, diagnostics: [] };
-					fileCache[currentRuleId].diagnostics.push({
-						...error,
-						file: undefined as any,
-						relatedInformation: error.relatedInformation?.map(info => ({
-							...info,
-							file: info.file ? { fileName: info.file.fileName } as any : undefined,
-						})),
-					});
-				}
 
 				let lintResult = lintResults.get(fileName);
 				if (!lintResult) {
@@ -273,6 +213,15 @@ export function createLinter(
 				}
 			}
 			return false;
+		},
+		// Per-diagnostic fix presence — used by the CLI cache layer to
+		// snapshot `hasFix` for a rule's cache entry without reaching
+		// into core's internal `diagnostic2Fixes` map.
+		hasFixForDiagnostic(fileName: string, diagnostic: ts.DiagnosticWithLocation): boolean {
+			const lintResult = lintResults.get(fileName);
+			if (!lintResult) return false;
+			const fixes = lintResult[1].get(diagnostic);
+			return !!fixes && fixes.length > 0;
 		},
 		getCodeFixes(
 			fileName: string,
